pyproject.toml
[tool.poetry]
name = "talkgpt"
version = "0.1.0"
description = "Talk to chatgpt"
authors = ["WincerChan <WincerChan@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.11"
edge-tts = "^6.1.3"
openai = "^0.27.1"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
ipython = "^8.11.0"
black = "^23.1.0"
PyAudio = "^0.2.13"
pydub = "^0.25.1"


[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "ipython<9.0.0,>=8.11.0",
    "black<24.0.0,>=23.1.0",
    "PyAudio<1.0.0,>=0.2.13",
    "pydub<1.0.0,>=0.25.1",
]

[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[project]
name = "talkgpt"
version = "0.1.0"
description = "Talk to chatgpt"
authors = [
    {name = "Wincer", email = "WincerChan+github@gmail.com"},
    {name = "WincerChan",email = "WincerChan@gmail.com"},
]
dependencies = [
    "edge-tts<7.0.0,>=6.1.3",
    "openai<1.0.0,>=0.27.1",
    "python-dotenv<2.0.0,>=1.0.0",
    "pyttsx3>=2.90",
    "pyobjc>=9.0.1",
    "gtts>=2.3.1",
    "pocketsphinx>=5.0.0",
    "SpeechRecognition>=3.9.0",
    "torch>=1.13.1",
]
requires-python = ">=3.11,<4.0"
license = {text = "GPLv3"}

[project.urls]
